refactor() Moving classes and removing legacy files
