--- src/database/schemas.py
@@ -1,17 +1,0 @@
-class Database:
-    """
-    Represents a database connection, session management, and associated SQLAlchemy Base.
-    Provides a method to create all tables for its Base.
-    """
-
-    def __init__(self, engine, session_maker, base):
-        self.engine = engine
-        self.session_maker = session_maker
-        self.base = base
-
-    def create_tables(self):
-        """Create all tables for the associated Base in this database."""
-        self.base.metadata.create_all(self.engine)
-
-    def __repr__(self):
-        return f"Database(engine={self.engine}, session_maker={self.session_maker}, base={self.base})"
